add filepath santization and overwrite prompt

+++ convert.py
@@ -5,7 +5,7 @@ import sys
 import sqlite3
 import xml.etree.ElementTree as ET
 from pathlib import Path
-
+from pathvalidate import sanitize_filepath, sanitize_filename
 
 from models.node import Node
 
@@ -15,42 +15,60 @@ def create_dir(n: Node, fid: int):
     pass
 
 def main():
-    ct_file = Path('testing').joinpath('ct_db.ctb')
+
+    print("starting...")
+    # check if arg size is correct
+    if (args_count := len(sys.argv)) == 2:
+        if sys.argv[1] == "help":
+            print("convert.py /PATH_TO_CHERRYTREE_DB/ /PATH_TO_NEW_DIR/")
+            raise SystemExit(0)
+    elif args_count > 3:
+        print("Too many args: type 'help' for usage")
+        raise SystemExit(2)
+    elif args_count < 3:
+        print("Too few args: type 'help' for usage")
+        raise SystemExit(2)
+
+
+    ct_file = Path(sys.argv[1])
+    target_dir = Path(sys.argv[2])
+
+    if target_dir.exists():
+        print(f'Directory: {target_dir} already exists!')
+        ans = input('overwrite? (y/n): \n')
+        if ans == 'y' or ans == 'yes':
+            shutil.rmtree(target_dir)
+        else:
+            sys.exit(1)
+
+    #troubleshooting path name
+    # p = Path('C:\\Users\\chase.clark\\vault\\cherrytree2markdown\\temp\\notes\\prn')
+    # print(len(p.as_posix()))
+    # p.mkdir(parents=True)
+
+    # ct_file = Path('testing').joinpath('ct_db.ctb')
     connection = sqlite3.connect(ct_file)
     cursor = connection.cursor()
     nodes = cursor.execute("SELECT * FROM node INNER JOIN children ON node.node_id = children.node_id ORDER BY father_id").fetchall()
-    print("starting...")
-    # if (args_count := len(sys.argv)) == 2:
-    #     if sys.argv[1] == "help":
-    #         print("convert.py /PATH_TO_CHERRYTREE_DB/ /PATH_TO_NEW_DIR/")
-    #         raise SystemExit(0)
-    # elif args_count > 3:
-    #     print("Too many args: type 'help' for usage")
-    #     raise SystemExit(2)
-    # elif args_count < 3:
-    #     print("Too few args: type 'help' for usage")
-    #     raise SystemExit(2)
-
-    # ct_file = Path(sys.argv[1])
-    # target_dir = Path(sys.argv[2])
     
-    # print(f'ct_file:{ct_file} target_dir:{target_dir}')
+    print(f'ct_file:{ct_file} target_dir:{target_dir}')
 
     # if target_dir.is_dir():
     #     print("The target directory exists already! this program will not overwrite! exiting...")
     #     raise SystemExit(1)
 
-    # temp hardcoding of inputs
-    target_dir = Path('temp')
-    if target_dir.exists():
-        shutil.rmtree(target_dir)
-
-
-    target_dir.mkdir() # create temp dir
+    # # temp hardcoding of inputs
+    # target_dir = Path('temp')
+    # if target_dir.exists():
+        
+    target_dir.mkdir() 
 
     for node in nodes:
         # take nodes from db and insert them into the dict
-        n = Node(node[0],node[1],node[2],node[6],node[7],node[8],node[14])
+        # replace troublesome chars
+        name = node[1].replace('/','(forward_slash)').replace('\\','(backslash)')
+        name = sanitize_filename(name)
+        n = Node(node[0],name,node[2],node[6],node[7],node[8],node[14])
         node_dict[n.id] = n
 
     # second pass through the nodes to make correct folder structure
@@ -68,12 +86,17 @@ def main():
         else:
             # get father node
             fn = node_dict[f]
-            path = fn.path.joinpath(n.name)
+            name = n.name
+            name = sanitize_filepath(name)
+            # print(sanitize_filepath(name,replacement_text=''))
+
+            path = fn.path.joinpath(name)
             if path.exists():
                 # generate new name for folder and node with duplicate name
-                new_name = f'{n.name}(dup)'
+                new_name = f'{name}(dup)'
                 path = fn.path.joinpath(new_name)
                 n.name = new_name
+            # print(f'path {path} node: {n}')
             path.mkdir()
             n.path = path
             
@@ -95,7 +118,7 @@ def main():
                         
                         # f.write(md.translate_xml(child.attrib,child.text,node_dict))
     connection.close()
-    print('finished')
+    print('finished with no errors')
 
 if __name__ == "__main__":
     main()
